Clear modal fields every time
