+++ core-ui/src/components/Lambdas/LambdaDetails/Tabs/Code/LambdaVariables/VariableForm/VariableModal.js
@@ -3,51 +3,8 @@ import React, { useState } from 'react';
 import { ModalWithForm } from 'react-shared';
 
 import VariableForm, { FORM_TYPE } from '../VariableForm/VariableForm';
-import {
-  VARIABLE_TYPE,
-  newVariableModel,
-} from 'components/Lambdas/helpers/lambdaVariables';
+import { VARIABLE_TYPE } from 'components/Lambdas/helpers/lambdaVariables';
 
-const EMPTY_VARIABLE_CUSTOM = {
-  type: VARIABLE_TYPE.CUSTOM,
-  variable: {
-    name: '',
-    value: '',
-  },
-  additionalProps: { dirty: true },
-};
-const EMPTY_VARIABLE_SECRET = {
-  type: VARIABLE_TYPE.SECRET,
-  variable: {
-    name: '',
-    valueFrom: {
-      secretKeyRef: {
-        name: null,
-        key: null,
-      },
-    },
-  },
-  additionalProps: { dirty: true },
-};
-
-const EMPTY_VARIABLE_CONFIG_MAP = {
-  type: VARIABLE_TYPE.CONFIG_MAP,
-  variable: {
-    name: '',
-    valueFrom: {
-      configMapKeyRef: {
-        name: null,
-        key: null,
-      },
-    },
-  },
-  additionalProps: { dirty: true },
-};
-const EMPTY_VARIABLE = {
-  CUSTOM: EMPTY_VARIABLE_CUSTOM,
-  SECRET: EMPTY_VARIABLE_SECRET,
-  CONFIG_MAP: EMPTY_VARIABLE_CONFIG_MAP,
-};
 export default function VariableModal({
   lambda,
   resources = [],
@@ -61,9 +18,6 @@ export default function VariableModal({
 }) {
   const [invalidModalPopupMessage, setInvalidModalPopupMessage] = useState('');
 
-  const [currentVariable, setCurrentVariable] = useState(
-    variable || newVariableModel(EMPTY_VARIABLE[type]),
-  );
   return (
     <ModalWithForm
       title={title}
@@ -77,8 +31,8 @@ export default function VariableModal({
           {...props}
           lambda={lambda}
           resources={resources}
-          currentVariable={currentVariable}
-          setCurrentVariable={setCurrentVariable}
+          variable={variable}
+          type={type}
           customVariables={customVariables}
           customValueFromVariables={customValueFromVariables}
           setInvalidModalPopupMessage={setInvalidModalPopupMessage}
